feat: add hide/unhide server functionality in UI settings
This commit introduces the ability for users to hide or unhide Plex servers. The hidden servers can be managed from the UI settings. New translations have been added to support this feature in different languages. The serverStore has been updated to include a getter for hidden servers. A new Vue component 'UnhideServersSection' has been created to handle the unhide server functionality in the settings UI.